Split observability replication status tests
