+++ .pm/tasks/task_bdb48338fac544849d8c681e9a7dd441.execution.md
@@ -484,3 +484,12 @@ Example:
 - Expected Result: if the repo helper exists, use it; otherwise create the PR through GitHub CLI with the already-verified `PR.md` packet.
 - Actual Result: `prepare-task-pr.sh` is absent; no existing PR was found for the branch; worktree was clean before recording this fallback.
 - Blocker / Next Action: commit this fallback note, create PR with `gh pr create`, then watch required checks/comments/mergeability.
+
+## 2026-06-23 20:42:00 CST / tpm
+- 完成内容: Fixed PR #581 required-gate file-size failure after merging latest `origin/main`.
+- 遗留事项: existing `oasis7_node` unused/dead-code warnings remain unrelated to this task.
+- Action: Split `build_chain_status_payload_zeroes_replication_gap_when_replication_disabled` out of `oasis7_chain_runtime_observability_tests.rs` into `oasis7_chain_runtime_observability_replication_tests.rs` and registered the new test module.
+- Validation Command: `env -u RUSTC_WRAPPER cargo fmt --check`; `./scripts/check-rust-file-size.sh`; `env -u RUSTC_WRAPPER cargo test -p oasis7 --bin oasis7_chain_runtime observability_tests -- --nocapture`; `env -u RUSTC_WRAPPER cargo test -p oasis7 --bin oasis7_chain_runtime observability_replication_tests -- --nocapture`; `env -u RUSTC_WRAPPER cargo test -p oasis7 --bin oasis7_chain_runtime observability_transport_tests -- --nocapture`; `./scripts/pm/workflow-lint.sh --task-uid task_bdb48338fac544849d8c681e9a7dd441 --phase pr-ready --allow-unbound`; `git diff --check`.
+- Expected Result: CI `check-rust-file-size.sh` blocker is removed without changing runtime behavior.
+- Actual Result: file-size check passed with zero oversized files; observability tests passed as 8 + 1 + 5 tests; PR-ready lint and diff check passed.
+- Blocker / Next Action: commit/push the CI fix and continue PR checks/comments/merge watch.
